Add SMSG_LOGOUT_CANCEL_ACK and SMSG_LOGOUT_COMPLETE for TBC/Wrath
