+++ wow_message_parser/wowm/world/login_logout/smsg_logout_cancel_ack.wowm
@@ -1,4 +1,4 @@
-#tag_all versions "1.12";
+#tag_all versions "1.12 2 3";
 
 smsg SMSG_LOGOUT_CANCEL_ACK = 0x4F { }
