+++ wow_message_parser/wowm/world/login_logout/smsg_logout_complete.wowm
@@ -1,4 +1,4 @@
-#tag_all versions "1.12";
+#tag_all versions "1.12 2 3";
 
 smsg SMSG_LOGOUT_COMPLETE = 0x4D { } {
     description = "Immediately logs out the client of the world and makes it send [CMSG_CHAR_ENUM].";
