+++ wow_world_messages/src/world/tbc/opcodes.rs
@@ -1705,6 +1705,8 @@ use crate::world::tbc::SMSG_NEW_WORLD;
 use crate::world::tbc::SMSG_TRANSFER_PENDING;
 use crate::world::tbc::SMSG_CHARACTER_LOGIN_FAILED;
 use crate::world::tbc::SMSG_LOGIN_SETTIMESPEED;
+use crate::world::tbc::SMSG_LOGOUT_COMPLETE;
+use crate::world::tbc::SMSG_LOGOUT_CANCEL_ACK;
 use crate::world::tbc::SMSG_CHANNEL_NOTIFY;
 use crate::world::tbc::SMSG_CHANNEL_LIST;
 use crate::world::tbc::SMSG_UPDATE_OBJECT;
@@ -1784,6 +1786,8 @@ pub enum ServerOpcodeMessage {
     SMSG_TRANSFER_PENDING(SMSG_TRANSFER_PENDING),
     SMSG_CHARACTER_LOGIN_FAILED(SMSG_CHARACTER_LOGIN_FAILED),
     SMSG_LOGIN_SETTIMESPEED(SMSG_LOGIN_SETTIMESPEED),
+    SMSG_LOGOUT_COMPLETE(SMSG_LOGOUT_COMPLETE),
+    SMSG_LOGOUT_CANCEL_ACK(SMSG_LOGOUT_CANCEL_ACK),
     SMSG_CHANNEL_NOTIFY(SMSG_CHANNEL_NOTIFY),
     SMSG_CHANNEL_LIST(SMSG_CHANNEL_LIST),
     SMSG_UPDATE_OBJECT(SMSG_UPDATE_OBJECT),
@@ -1865,6 +1869,8 @@ impl ServerOpcodeMessage {
             0x003F => Ok(Self::SMSG_TRANSFER_PENDING(<SMSG_TRANSFER_PENDING as crate::Message>::read_body(&mut r, body_size)?)),
             0x0041 => Ok(Self::SMSG_CHARACTER_LOGIN_FAILED(<SMSG_CHARACTER_LOGIN_FAILED as crate::Message>::read_body(&mut r, body_size)?)),
             0x0042 => Ok(Self::SMSG_LOGIN_SETTIMESPEED(<SMSG_LOGIN_SETTIMESPEED as crate::Message>::read_body(&mut r, body_size)?)),
+            0x004D => Ok(Self::SMSG_LOGOUT_COMPLETE(<SMSG_LOGOUT_COMPLETE as crate::Message>::read_body(&mut r, body_size)?)),
+            0x004F => Ok(Self::SMSG_LOGOUT_CANCEL_ACK(<SMSG_LOGOUT_CANCEL_ACK as crate::Message>::read_body(&mut r, body_size)?)),
             0x0099 => Ok(Self::SMSG_CHANNEL_NOTIFY(<SMSG_CHANNEL_NOTIFY as crate::Message>::read_body(&mut r, body_size)?)),
             0x009B => Ok(Self::SMSG_CHANNEL_LIST(<SMSG_CHANNEL_LIST as crate::Message>::read_body(&mut r, body_size)?)),
             0x00A9 => Ok(Self::SMSG_UPDATE_OBJECT(<SMSG_UPDATE_OBJECT as crate::Message>::read_body(&mut r, body_size)?)),
@@ -2014,6 +2020,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_TRANSFER_PENDING(c) => c.write_encrypted_server(w, e),
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.write_encrypted_server(w, e),
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.write_encrypted_server(w, e),
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.write_encrypted_server(w, e),
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.write_encrypted_server(w, e),
             Self::SMSG_CHANNEL_NOTIFY(c) => c.write_encrypted_server(w, e),
             Self::SMSG_CHANNEL_LIST(c) => c.write_encrypted_server(w, e),
             Self::SMSG_UPDATE_OBJECT(c) => c.write_encrypted_server(w, e),
@@ -2096,6 +2104,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_TRANSFER_PENDING(c) => c.write_unencrypted_server(w),
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.write_unencrypted_server(w),
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.write_unencrypted_server(w),
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.write_unencrypted_server(w),
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.write_unencrypted_server(w),
             Self::SMSG_CHANNEL_NOTIFY(c) => c.write_unencrypted_server(w),
             Self::SMSG_CHANNEL_LIST(c) => c.write_unencrypted_server(w),
             Self::SMSG_UPDATE_OBJECT(c) => c.write_unencrypted_server(w),
@@ -2178,6 +2188,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_TRANSFER_PENDING(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.tokio_write_encrypted_server(w, e).await,
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.tokio_write_encrypted_server(w, e).await,
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_CHANNEL_NOTIFY(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_CHANNEL_LIST(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_UPDATE_OBJECT(c) => c.tokio_write_encrypted_server(w, e).await,
@@ -2260,6 +2272,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_TRANSFER_PENDING(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.tokio_write_unencrypted_server(w).await,
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.tokio_write_unencrypted_server(w).await,
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_CHANNEL_NOTIFY(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_CHANNEL_LIST(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_UPDATE_OBJECT(c) => c.tokio_write_unencrypted_server(w).await,
@@ -2342,6 +2356,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_TRANSFER_PENDING(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.astd_write_encrypted_server(w, e).await,
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.astd_write_encrypted_server(w, e).await,
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_CHANNEL_NOTIFY(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_CHANNEL_LIST(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_UPDATE_OBJECT(c) => c.astd_write_encrypted_server(w, e).await,
@@ -2424,6 +2440,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_TRANSFER_PENDING(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.astd_write_unencrypted_server(w).await,
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.astd_write_unencrypted_server(w).await,
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_CHANNEL_NOTIFY(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_CHANNEL_LIST(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_UPDATE_OBJECT(c) => c.astd_write_unencrypted_server(w).await,
@@ -2508,6 +2526,8 @@ impl std::fmt::Display for ServerOpcodeMessage {
             ServerOpcodeMessage::SMSG_TRANSFER_PENDING(_) => "SMSG_TRANSFER_PENDING",
             ServerOpcodeMessage::SMSG_CHARACTER_LOGIN_FAILED(_) => "SMSG_CHARACTER_LOGIN_FAILED",
             ServerOpcodeMessage::SMSG_LOGIN_SETTIMESPEED(_) => "SMSG_LOGIN_SETTIMESPEED",
+            ServerOpcodeMessage::SMSG_LOGOUT_COMPLETE(_) => "SMSG_LOGOUT_COMPLETE",
+            ServerOpcodeMessage::SMSG_LOGOUT_CANCEL_ACK(_) => "SMSG_LOGOUT_CANCEL_ACK",
             ServerOpcodeMessage::SMSG_CHANNEL_NOTIFY(_) => "SMSG_CHANNEL_NOTIFY",
             ServerOpcodeMessage::SMSG_CHANNEL_LIST(_) => "SMSG_CHANNEL_LIST",
             ServerOpcodeMessage::SMSG_UPDATE_OBJECT(_) => "SMSG_UPDATE_OBJECT",
@@ -2625,6 +2645,18 @@ impl From<SMSG_CHARACTER_LOGIN_FAILED> for ServerOpcodeMessage {
 impl From<SMSG_LOGIN_SETTIMESPEED> for ServerOpcodeMessage {
     fn from(c: SMSG_LOGIN_SETTIMESPEED) -> Self {
         Self::SMSG_LOGIN_SETTIMESPEED(c)
+    }
+}
+
+impl From<SMSG_LOGOUT_COMPLETE> for ServerOpcodeMessage {
+    fn from(c: SMSG_LOGOUT_COMPLETE) -> Self {
+        Self::SMSG_LOGOUT_COMPLETE(c)
+    }
+}
+
+impl From<SMSG_LOGOUT_CANCEL_ACK> for ServerOpcodeMessage {
+    fn from(c: SMSG_LOGOUT_CANCEL_ACK) -> Self {
+        Self::SMSG_LOGOUT_CANCEL_ACK(c)
     }
 }
