+++ wow_world_messages/src/world/tbc/smsg_logout_complete.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::smsg_logout_complete1_1223::SMSG_LOGOUT_COMPLETE;
+
