+++ wow_world_messages/src/world/wrath/mod.rs
@@ -394,6 +394,10 @@ pub(crate) mod smsg_login_settimespeed;
 pub use smsg_login_settimespeed::*;
 pub(crate) mod smsg_login_verify_world;
 pub use smsg_login_verify_world::*;
+pub(crate) mod smsg_logout_cancel_ack;
+pub use smsg_logout_cancel_ack::*;
+pub(crate) mod smsg_logout_complete;
+pub use smsg_logout_complete::*;
 pub(crate) mod smsg_logout_response;
 pub use smsg_logout_response::*;
 pub(crate) mod smsg_new_world;
