+++ wow_world_messages/src/world/wrath/opcodes.rs
@@ -1730,6 +1730,8 @@ use crate::world::wrath::SMSG_TRANSFER_PENDING;
 use crate::world::wrath::SMSG_CHARACTER_LOGIN_FAILED;
 use crate::world::wrath::SMSG_LOGIN_SETTIMESPEED;
 use crate::world::wrath::SMSG_LOGOUT_RESPONSE;
+use crate::world::wrath::SMSG_LOGOUT_COMPLETE;
+use crate::world::wrath::SMSG_LOGOUT_CANCEL_ACK;
 use crate::world::wrath::SMSG_CONTACT_LIST;
 use crate::world::wrath::SMSG_CHANNEL_NOTIFY;
 use crate::world::wrath::SMSG_CHANNEL_LIST;
@@ -1821,6 +1823,8 @@ pub enum ServerOpcodeMessage {
     SMSG_CHARACTER_LOGIN_FAILED(SMSG_CHARACTER_LOGIN_FAILED),
     SMSG_LOGIN_SETTIMESPEED(SMSG_LOGIN_SETTIMESPEED),
     SMSG_LOGOUT_RESPONSE(SMSG_LOGOUT_RESPONSE),
+    SMSG_LOGOUT_COMPLETE(SMSG_LOGOUT_COMPLETE),
+    SMSG_LOGOUT_CANCEL_ACK(SMSG_LOGOUT_CANCEL_ACK),
     SMSG_CONTACT_LIST(SMSG_CONTACT_LIST),
     SMSG_CHANNEL_NOTIFY(SMSG_CHANNEL_NOTIFY),
     SMSG_CHANNEL_LIST(SMSG_CHANNEL_LIST),
@@ -1914,6 +1918,8 @@ impl ServerOpcodeMessage {
             0x0041 => Ok(Self::SMSG_CHARACTER_LOGIN_FAILED(<SMSG_CHARACTER_LOGIN_FAILED as crate::Message>::read_body(&mut r, body_size)?)),
             0x0042 => Ok(Self::SMSG_LOGIN_SETTIMESPEED(<SMSG_LOGIN_SETTIMESPEED as crate::Message>::read_body(&mut r, body_size)?)),
             0x004C => Ok(Self::SMSG_LOGOUT_RESPONSE(<SMSG_LOGOUT_RESPONSE as crate::Message>::read_body(&mut r, body_size)?)),
+            0x004D => Ok(Self::SMSG_LOGOUT_COMPLETE(<SMSG_LOGOUT_COMPLETE as crate::Message>::read_body(&mut r, body_size)?)),
+            0x004F => Ok(Self::SMSG_LOGOUT_CANCEL_ACK(<SMSG_LOGOUT_CANCEL_ACK as crate::Message>::read_body(&mut r, body_size)?)),
             0x0067 => Ok(Self::SMSG_CONTACT_LIST(<SMSG_CONTACT_LIST as crate::Message>::read_body(&mut r, body_size)?)),
             0x0099 => Ok(Self::SMSG_CHANNEL_NOTIFY(<SMSG_CHANNEL_NOTIFY as crate::Message>::read_body(&mut r, body_size)?)),
             0x009B => Ok(Self::SMSG_CHANNEL_LIST(<SMSG_CHANNEL_LIST as crate::Message>::read_body(&mut r, body_size)?)),
@@ -2156,6 +2162,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.write_encrypted_server(w, e),
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.write_encrypted_server(w, e),
             Self::SMSG_LOGOUT_RESPONSE(c) => c.write_encrypted_server(w, e),
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.write_encrypted_server(w, e),
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.write_encrypted_server(w, e),
             Self::SMSG_CONTACT_LIST(c) => c.write_encrypted_server(w, e),
             Self::SMSG_CHANNEL_NOTIFY(c) => c.write_encrypted_server(w, e),
             Self::SMSG_CHANNEL_LIST(c) => c.write_encrypted_server(w, e),
@@ -2250,6 +2258,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.write_unencrypted_server(w),
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.write_unencrypted_server(w),
             Self::SMSG_LOGOUT_RESPONSE(c) => c.write_unencrypted_server(w),
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.write_unencrypted_server(w),
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.write_unencrypted_server(w),
             Self::SMSG_CONTACT_LIST(c) => c.write_unencrypted_server(w),
             Self::SMSG_CHANNEL_NOTIFY(c) => c.write_unencrypted_server(w),
             Self::SMSG_CHANNEL_LIST(c) => c.write_unencrypted_server(w),
@@ -2344,6 +2354,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_LOGOUT_RESPONSE(c) => c.tokio_write_encrypted_server(w, e).await,
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.tokio_write_encrypted_server(w, e).await,
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_CONTACT_LIST(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_CHANNEL_NOTIFY(c) => c.tokio_write_encrypted_server(w, e).await,
             Self::SMSG_CHANNEL_LIST(c) => c.tokio_write_encrypted_server(w, e).await,
@@ -2438,6 +2450,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_LOGOUT_RESPONSE(c) => c.tokio_write_unencrypted_server(w).await,
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.tokio_write_unencrypted_server(w).await,
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_CONTACT_LIST(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_CHANNEL_NOTIFY(c) => c.tokio_write_unencrypted_server(w).await,
             Self::SMSG_CHANNEL_LIST(c) => c.tokio_write_unencrypted_server(w).await,
@@ -2532,6 +2546,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_LOGOUT_RESPONSE(c) => c.astd_write_encrypted_server(w, e).await,
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.astd_write_encrypted_server(w, e).await,
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_CONTACT_LIST(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_CHANNEL_NOTIFY(c) => c.astd_write_encrypted_server(w, e).await,
             Self::SMSG_CHANNEL_LIST(c) => c.astd_write_encrypted_server(w, e).await,
@@ -2626,6 +2642,8 @@ impl ServerOpcodeMessage {
             Self::SMSG_CHARACTER_LOGIN_FAILED(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_LOGIN_SETTIMESPEED(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_LOGOUT_RESPONSE(c) => c.astd_write_unencrypted_server(w).await,
+            Self::SMSG_LOGOUT_COMPLETE(c) => c.astd_write_unencrypted_server(w).await,
+            Self::SMSG_LOGOUT_CANCEL_ACK(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_CONTACT_LIST(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_CHANNEL_NOTIFY(c) => c.astd_write_unencrypted_server(w).await,
             Self::SMSG_CHANNEL_LIST(c) => c.astd_write_unencrypted_server(w).await,
@@ -2722,6 +2740,8 @@ impl std::fmt::Display for ServerOpcodeMessage {
             ServerOpcodeMessage::SMSG_CHARACTER_LOGIN_FAILED(_) => "SMSG_CHARACTER_LOGIN_FAILED",
             ServerOpcodeMessage::SMSG_LOGIN_SETTIMESPEED(_) => "SMSG_LOGIN_SETTIMESPEED",
             ServerOpcodeMessage::SMSG_LOGOUT_RESPONSE(_) => "SMSG_LOGOUT_RESPONSE",
+            ServerOpcodeMessage::SMSG_LOGOUT_COMPLETE(_) => "SMSG_LOGOUT_COMPLETE",
+            ServerOpcodeMessage::SMSG_LOGOUT_CANCEL_ACK(_) => "SMSG_LOGOUT_CANCEL_ACK",
             ServerOpcodeMessage::SMSG_CONTACT_LIST(_) => "SMSG_CONTACT_LIST",
             ServerOpcodeMessage::SMSG_CHANNEL_NOTIFY(_) => "SMSG_CHANNEL_NOTIFY",
             ServerOpcodeMessage::SMSG_CHANNEL_LIST(_) => "SMSG_CHANNEL_LIST",
@@ -2981,6 +3001,18 @@ impl From<SMSG_LOGIN_SETTIMESPEED> for ServerOpcodeMessage {
 impl From<SMSG_LOGOUT_RESPONSE> for ServerOpcodeMessage {
     fn from(c: SMSG_LOGOUT_RESPONSE) -> Self {
         Self::SMSG_LOGOUT_RESPONSE(c)
+    }
+}
+
+impl From<SMSG_LOGOUT_COMPLETE> for ServerOpcodeMessage {
+    fn from(c: SMSG_LOGOUT_COMPLETE) -> Self {
+        Self::SMSG_LOGOUT_COMPLETE(c)
+    }
+}
+
+impl From<SMSG_LOGOUT_CANCEL_ACK> for ServerOpcodeMessage {
+    fn from(c: SMSG_LOGOUT_CANCEL_ACK) -> Self {
+        Self::SMSG_LOGOUT_CANCEL_ACK(c)
     }
 }
