+++ wow_world_messages/src/world/wrath/smsg_logout_cancel_ack.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::smsg_logout_cancel_ack1_1223::SMSG_LOGOUT_CANCEL_ACK;
+
