+++ wowm_language/src/docs/smsg_logout_cancel_ack.md
@@ -1,6 +1,6 @@
 # SMSG_LOGOUT_CANCEL_ACK
 
-## Client Version 1.12
+## Client Version 1.12, Client Version 2, Client Version 3
 
 ### Wowm Representation
