+++ wowm_language/src/docs/smsg_logout_complete.md
@@ -1,6 +1,6 @@
 # SMSG_LOGOUT_COMPLETE
 
-## Client Version 1.12
+## Client Version 1.12, Client Version 2, Client Version 3
 
 ### Description
